requirements.txt
slack-bolt==1.20.1
slack-sdk==3.31.0
pykrx==1.0.45
opendartreader==0.2.1
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
